pr-reporter/requirements.txt
PyGithub==1.59.0
openpyxl==3.1.2
